Moved network-related classes to own folder Signed-off-by:doe300 <stadeldani@web.de>
