+++ src/sip/STUNClient.cpp
@@ -6,7 +6,7 @@
  */
 
 #include "sip/STUNClient.h"
-#include "UDPWrapper.h"
+#include "network/UDPWrapper.h"
 
 #include <chrono>
